VPC VTEP needed in Overlay info
First time doing this.   I had a customer looking to use ACITOOLKIT to
discover where every endpoint was.  The TEP to Endpoint was not easy
via GUI so ACITOOLKIT with -overlay gave great information.  What it
did not do was provide the VPC VTEP info which is where the Tunnels
source and terminate.  The current info only provides a switches
specific VTEP not a shared VPC VTEP.